requirements.txt
# ===========================================
# Core
# ===========================================
pydantic>=2.10.6
pydantic-settings>=2.7.1
python-dotenv>=1.0.1

# ===========================================
# PDF Processing
# ===========================================
pdfplumber>=0.11.7
pdf2image>=1.17.0
pytesseract>=0.3.13
opencv-python>=4.10.0

# ===========================================
# LangChain Ecosystem (v0.3.x stable)
# ===========================================
langchain>=0.3.18
langchain-community>=0.3.18
langchain-core>=0.3.28
langchain-text-splitters>=0.3.5
langchain-openai>=0.3.0

# ===========================================
# Embeddings & ML
# ===========================================
sentence-transformers>=3.3.1
torch>=2.5.1  # Explicit install recommended
transformers>=4.47.1

# ===========================================
# Vector Stores
# ===========================================
chromadb>=0.6.3
# faiss-cpu>=1.9.0  # Uncomment if needed

# ===========================================
# LLM Integration
# ===========================================
openai>=1.59.5
tiktoken>=0.8.0

# ===========================================
# Utilities
# ===========================================
tqdm>=4.67.1
numpy>=2.2.6
protobuf>=5.26.1

# ===========================================
# Testing
# ===========================================
pytest>=8.4.2
pytest-cov>=6.0.0